requirements.txt
SQLAlchemy==1.2.5
pymysql
python-telegram-bot
